策略权重装配处理

+++ big_market-infrastructure/src/main/java/org/example/infrastructure/persistent/po/StrategyRule.java
@@ -10,7 +10,7 @@ public class StrategyRule {
     private String id;
 
     /**  抽奖策略ID  */
-    private Integer strategyId;
+    private Long strategyId;
 
     /**  抽奖奖品ID  */
     private Integer awardId;
